+++ urid/lv2-urid/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "lv2-urid"
+version = "1.0.1"
+authors = ["Jan-Oliver 'Janonard' Opdenhövel <jan.opdenhoevel@protonmail.com>", "Adrien Prokopowicz <adrien.prokopowicz@gmail.com>"]
+edition = "2018"
+license = "MIT OR Apache-2.0"
+
+description = "rust-lv2's URID handling library"
+readme = "README.md"
+repository = "https://github.com/RustAudio/rust-lv2"
+
+[badges]
+travis-ci = { repository = "RustAudio/rust-lv2", branch = "master" }
+maintenance = { status = "passively-maintained" }
+
+[dependencies]
+lv2-core = "1.0.0"
+lv2-sys = "1.0.0"
+urid = "0.1.0"
