+++ urid/lv2-urid/src/lib.rs
@@ -1,0 +1,14 @@
+//! LV2 specification centered around the Map feature.
+//!
+//! The URID specification provides a host feature that can be used by plugins to map URIs to integers, so-called URIDs. These URIDs are used by many other specifications to identify other URI bounds and combine the flexibility of URIs with the comparison speed of integers.
+//!
+//! Since this crate depends on `-sys` crates that use `bindgen` to create the C API bindings,
+//! you need to have clang installed on your machine.
+extern crate lv2_core as core;
+extern crate lv2_sys as sys;
+
+mod feature;
+mod mapper;
+
+pub use feature::*;
+pub use mapper::*;
