+++ urid/tests/urid.rs
@@ -1,5 +1,3 @@
-use lv2_urid::*;
-use std::pin::Pin;
 use urid::*;
 
 struct MyTypeA;
@@ -16,9 +14,7 @@ unsafe impl UriBound for MyTypeB {
 
 #[test]
 fn test_map() {
-    let mut host_map: Pin<Box<HostMap<HashURIDMapper>>> = Box::pin(HashURIDMapper::new().into());
-    let map_interface = host_map.as_mut().make_map_interface();
-    let map = LV2Map::new(&map_interface);
+    let map = HashURIDMapper::new();
 
     assert_eq!(1, map.map_uri(MyTypeA::uri()).unwrap());
     assert_eq!(1, map.map_type::<MyTypeA>().unwrap());
@@ -32,11 +28,7 @@ fn test_map() {
 
 #[test]
 fn test_unmap() {
-    let mut host_map: Pin<Box<HostMap<HashURIDMapper>>> = Box::pin(HashURIDMapper::new().into());
-    let map_interface = host_map.as_mut().make_map_interface();
-    let map = LV2Map::new(&map_interface);
-    let unmap_interface = host_map.as_mut().make_unmap_interface();
-    let unmap = LV2Unmap::new(&unmap_interface);
+    let map = HashURIDMapper::new();
 
     let (type_a, type_b) = {
         (
@@ -45,8 +37,8 @@ fn test_unmap() {
         )
     };
 
-    assert_eq!(MyTypeA::uri(), unmap.unmap(type_a).unwrap());
-    assert_eq!(MyTypeB::uri(), unmap.unmap(type_b).unwrap());
+    assert_eq!(MyTypeA::uri(), map.unmap(type_a).unwrap());
+    assert_eq!(MyTypeB::uri(), map.unmap(type_b).unwrap());
 }
 
 #[derive(URIDCollection)]
@@ -57,9 +49,7 @@ struct MyURIDCollection {
 
 #[test]
 fn test_collection() {
-    let mut host_map: Pin<Box<HostMap<HashURIDMapper>>> = Box::pin(HashURIDMapper::new().into());
-    let map_interface = host_map.as_mut().make_map_interface();
-    let map = LV2Map::new(&map_interface);
+    let map = HashURIDMapper::new();
     let collection = MyURIDCollection::from_map(&map).unwrap();
 
     assert_eq!(1, collection.type_a);
